[add] 增加编码模块、msg队列以及merge部分实现 [add] 增加Common、config、log等公共模块

+++ src/CCatClient/ZRSafeQueue.h
@@ -37,7 +37,7 @@ void clearZRSafeQueue(ZRSafeQueue * pQueue);
 void destroyZRSafeQueue(ZRSafeQueue * pQueue);
 
 // 下面这些加锁也没啥用，干脆不加
-inline int getZRSafeQueueSize(ZRSafeQueue * pQueue)
+inline size_t getZRSafeQueueSize(ZRSafeQueue * pQueue)
 {
 	return getZRStaticQueueSize(pQueue->queue);
 }
